session attribute user changed to userId

+++ src/main/java/com/tms/controller/UserAction.java
@@ -10,7 +10,6 @@ import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.swing.text.html.parser.Entity;
 import java.io.IOException;
 import java.util.*;
 
@@ -18,22 +17,34 @@ public class UserAction {
     private static UserDao userDao = new UserDao();
     final static Logger logger = Logger.getLogger(UserAction.class);
 
-    public static void logIn(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp, User user) {
-        User tempUser = null;
+    public static void logIn(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp, Long id) {
+        Long userId = 0L;
+        if (id == 0) {
+            String email;
+            String password;
 
-        Optional optionalUser = userDao.get(user.getEmail(), user.getPassword());
-        if (optionalUser.isPresent()) {
-            tempUser = (User) optionalUser.get();
-            req.getSession().setAttribute("user", tempUser);
-            showInfoPage(servletContext, req, resp);
+            email = req.getParameter("email");
+            password = req.getParameter("password");
+
+            Optional<User> optionalUser = userDao.get(email, password);
+            if (optionalUser.isPresent()) {
+                userId = optionalUser.get().getId();
+            } else {
+                emulateError(servletContext, req, resp, "Указан неверный e-mail или пароль");
+            }
         } else {
-            emulateError(servletContext, req, resp, "Указан неверный e-mail или пароль");
+            userId = id;
         }
+        req.getSession().setAttribute("userId", userId);
+        showInfoPage(servletContext, req, resp);
     }
 
     public static void showInfoPage(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp) {
         List<User> userList;
-        User user = (User) req.getSession().getAttribute("user");
+
+        Optional<User> optionalUser = UserUtil.getSessionUser(req);
+
+        User user = optionalUser.get();
         req.setAttribute("user", user);
 
         Map<String, Boolean> allUserRoles = RoleUtil.getAllUserRoles(user);
@@ -41,9 +52,12 @@ public class UserAction {
 
         if (allUserRoles.get("show")) {
             userList = userDao.getAll();
-            req.setAttribute("userList", userList);
-            req.getSession().setAttribute("userList", userList);
+        } else {
+            userList = Arrays.asList(user);
         }
+
+        req.setAttribute("userList", userList);
+        req.getSession().setAttribute("userList", userList);
 
         try {
             servletContext.getRequestDispatcher("/userInfo.jsp").forward(req, resp);
@@ -53,7 +67,7 @@ public class UserAction {
     }
 
     public static void logOut(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp) {
-        req.getSession().setAttribute("user", null);
+        req.getSession().setAttribute("userID", 0);
         try {
             servletContext.getRequestDispatcher("/loginForm.jsp").forward(req, resp);
         } catch (ServletException | IOException e) {
@@ -62,38 +76,49 @@ public class UserAction {
     }
 
     public static void editUser(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp) {
-        long userId = Long.valueOf(req.getParameterValues("id")[0]);
-        User user = userDao.getRecordById(userId).get();
+        if (RoleUtil.isUserPlaysRole(UserUtil.getSessionUser(req).get(), "edit")) {
+            long userId = Long.valueOf(req.getParameterValues("id")[0]);
+            User user = userDao.getRecordById(userId).get();
 
-        req.setAttribute("user", user);
+            req.setAttribute("user", user);
 
-        Map<String, Boolean> getAllUserRoles = RoleUtil.getAllUserRoles(user);
-        getAllUserRoles.forEach((key, value) -> req.setAttribute(key, value));
+            Map<String, Boolean> getAllUserRoles = RoleUtil.getAllUserRoles(user);
+            getAllUserRoles.forEach((key, value) -> req.setAttribute(key, value));
 
-        try {
-            servletContext.getRequestDispatcher("/edit.jsp").forward(req, resp);
-        } catch (ServletException | IOException e) {
-            logger.error("error with open update page " + e.getMessage());
+            try {
+                servletContext.getRequestDispatcher("/edit.jsp").forward(req, resp);
+            } catch (ServletException | IOException e) {
+                logger.error("error with open update page " + e.getMessage());
+            }
+        } else {
+            emulateError(servletContext, req, resp, "Ошибка доступа");
         }
     }
 
     public static void updateUser(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp) {
-        UserUtil.updateUser(req);
-        showInfoPage(servletContext, req, resp);
+        if (RoleUtil.isUserPlaysRole(UserUtil.getSessionUser(req).get(), "edit")) {
+            UserUtil.updateUser(req);
+            showInfoPage(servletContext, req, resp);
+        } else {
+            emulateError(servletContext, req, resp, "Ошибка доступа");
+        }
     }
 
     public static void deleteUser(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp) {
         long userId = Long.valueOf(req.getParameterValues("id")[0]);
-        User user = userDao.getRecordById(userId).get();
 
-        User curUser = (User) req.getSession().getAttribute("user");
+        if (RoleUtil.isUserPlaysRole(UserUtil.getSessionUser(req).get(), "delete")) {
+            userDao.delete(userId);
 
-        userDao.delete(userId);
-
-        if (curUser.equals(user)) {
-            emulateError(servletContext, req, resp, "Вы удалили себя");
+            Long curUserId = UserUtil.getSessionUserId(req);
+            if (userId == curUserId) {
+                emulateError(servletContext, req, resp, "Вы удалили себя");
+                return;
+            }
+            showInfoPage(servletContext, req, resp);
+        } else {
+            emulateError(servletContext, req, resp, "Ошибка доступа");
         }
-        showInfoPage(servletContext, req, resp);
     }
 
     public static void emulateError(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp, String message) {
@@ -105,7 +130,7 @@ public class UserAction {
     public static void saveUser(ServletContext servletContext, HttpServletRequest req, HttpServletResponse resp) {
         User newUser = UserUtil.saveUser(req);
         if (newUser != null) {
-            UserAction.logIn(servletContext, req, resp, newUser);
+            UserAction.logIn(servletContext, req, resp, newUser.getId());
         } else {
             UserAction.emulateError(servletContext, req, resp, "Пользователь с таким e-mail уже существует");
         }
